+++ src/components/headQuarter/DashboardNewOrderCard.vue
